Renamed ConsultantMenu and -Layout to ContractorMenu and -Layout, also removed contractorId and ownerId from path
