+++ src/layout/AppSidebar.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ConsultantMenu from './ConsultantMenu.vue';
+import ContractorMenu from './ContractorMenu.vue';
 import ManagerMenu from './ManagerMenu.vue';
 import TenancyMenu from './TenancyMenu.vue';
 import { useProjectStore } from '@/stores/ProjectStore';
@@ -24,7 +24,7 @@ const projectStore = useProjectStore();
     "
     class="layout-sidebar"
   >
-    <ConsultantMenu></ConsultantMenu>
+    <ContractorMenu></ContractorMenu>
   </div>
   <div v-if="projectStore.selectedProject?.memberRole === 'LESSEE'" class="layout-sidebar">
     <TenancyMenu></TenancyMenu>
